CHeck llvm-link command
